reorganize codegen
