+++ codegen/src/io.rs
@@ -3,703 +3,11 @@ use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 use syn::{parse_quote, DeriveInput};
 
+use crate::common::*;
+use crate::layout::*;
+use crate::schematic::*;
+
 use crate::substrate_ident;
-
-fn impl_clone(helper: &DeriveInputHelper) -> TokenStream {
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(|ty, _| parse_quote! { #ty: ::std::clone::Clone });
-    let clone_body = helper.map_data(&parse_quote! { Self }, |MapField { ty, refer, .. }| {
-        quote! { <#ty as ::std::clone::Clone>::clone(#refer) }
-    });
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { ::std::clone::Clone },
-        trait_body: quote! {
-            fn clone(&self) -> Self {
-                #clone_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_debug(helper: &DeriveInputHelper) -> TokenStream {
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(|ty, _| parse_quote! { #ty: ::std::fmt::Debug });
-    let ident = helper.get_ident();
-    let debug_body = helper.map(|fields| {
-        let mapped_fields = fields.iter().map(
-            |MapField {
-                 refer,
-                 pretty_ident,
-                 ..
-             }| quote! { .field(stringify!(#pretty_ident), #refer) },
-        );
-        quote! { __substrate_f.debug_struct(stringify!(#ident))#(#mapped_fields)*.finish() }
-    });
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { ::std::fmt::Debug },
-        trait_body: quote! {
-            fn fmt(&self, __substrate_f: &mut ::std::fmt::Formatter<'_>) -> ::std::fmt::Result {
-                #debug_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_partial_eq(helper: &DeriveInputHelper) -> TokenStream {
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(|ty, _| parse_quote! { #ty: ::std::fmt::Debug });
-    let partial_eq_body = helper.double_map(
-        (&quote! { self }, &quote! { __substrate_other }),
-        |fields: &[(&MapField, &MapField)]| {
-            if fields.is_empty() {
-                quote! { true }
-            } else {
-                let mapped_fields = fields.iter().map(
-                    |(
-                        MapField {
-                            ty, refer: refer0, ..
-                        },
-                        MapField { refer: refer1, .. },
-                    )| quote! { <#ty as ::std::cmp::PartialEq>::eq(#refer0, #refer1) },
-                );
-                quote! { #(#mapped_fields)&&* }
-            }
-        },
-        quote! { false },
-    );
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { ::std::cmp::PartialEq },
-        trait_body: quote! {
-            fn eq(&self, __substrate_other: &Self) -> bool {
-                #partial_eq_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_eq(helper: &DeriveInputHelper) -> TokenStream {
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(|ty, _| parse_quote! { #ty: ::std::cmp::Eq });
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { ::std::cmp::Eq },
-        trait_body: quote! {},
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_flatlen(helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(|ty, _| parse_quote! { #ty: #substrate::types::FlatLen });
-    let flatlen_body = helper.map(|fields| {
-        if fields.is_empty() {
-            quote! { 0 }
-        } else {
-            let mapped_fields = fields
-                .iter()
-                .map(|MapField{ty, refer, ..}| quote! { <#ty as #substrate::types::FlatLen>::len(#refer) });
-            quote! { #(#mapped_fields)+* }
-        }
-    });
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::FlatLen },
-        trait_body: quote! {
-            fn len(&self) -> usize {
-                #flatlen_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_flatten_direction(helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::Flatten<#substrate::types::Direction> },
-    );
-    let flatten_direction_body = helper.map(
-            |fields| {
-                let mapped_fields = fields.iter().map(|MapField { ty, refer, .. }| quote! { <#ty as #substrate::types::Flatten<#substrate::types::Direction>>::flatten(#refer, __substrate_output_sink); });
-                quote! { #(#mapped_fields)* }
-            },
-        );
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::Flatten<#substrate::types::Direction> },
-        trait_body: quote! {
-            fn flatten<E>(&self, __substrate_output_sink: &mut E)
-            where
-                E: ::std::iter::Extend<#substrate::types::Direction> {
-                #flatten_direction_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_has_bundle_kind(
-    helper: &DeriveInputHelper,
-    kind_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::HasBundleKind },
-    );
-    let bundle_kind = kind_helper.get_type();
-    let kind_body = helper.map_data(
-        &parse_quote! { #bundle_kind },
-        |MapField { ty, refer, .. }| {
-            quote! { <#ty as #substrate::types::HasBundleKind>::kind(#refer) }
-        },
-    );
-    let bundle_kind_full = kind_helper.get_full_type();
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::HasBundleKind },
-        trait_body: quote! {
-            type BundleKind = #bundle_kind_full;
-
-            fn kind(&self) -> <Self as #substrate::types::HasBundleKind>::BundleKind {
-                #kind_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_view_source(helper: &DeriveInputHelper, other: Option<&syn::Type>) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::HasBundleKind },
-    );
-    let view_source_body = if let Some(other) = other {
-        quote! {
-            type Kind = #substrate::types::codegen::FromOther;
-            type Source = #other;
-        }
-    } else {
-        quote! {
-            type Kind = #substrate::types::codegen::FromSelf;
-            type Source = Self;
-        }
-    };
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::codegen::ViewSource },
-        trait_body: view_source_body,
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_has_name_tree(helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::HasBundleKind },
-    );
-    let name_fields = helper.map(|fields| {
-        let mapped_fields = fields.iter().map(|MapField{ ty, pretty_ident, refer, .. }|
-            quote! {
-                (#substrate::arcstr::literal!(::std::stringify!(#pretty_ident)), <#ty as #substrate::types::HasNameTree>::names(&#refer))
-            }
-        );
-        quote! { vec![ #(#mapped_fields),* ] }
-    });
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::HasNameTree },
-        trait_body: quote! {
-            fn names(&self) -> ::std::option::Option<::std::vec::Vec<#substrate::types::NameTree>> {
-                let v: ::std::vec::Vec<#substrate::types::NameTree> =  #name_fields
-                     .into_iter()
-                     .filter_map(|(frag, children)| children.map(|c| #substrate::types::NameTree::new(frag, c)))
-                     .collect();
-                if v.len() == 0 { None } else { Some(v) }
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_flatten_generic(helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let flatten_generic = parse_quote! { SubstrateF };
-    let mut helper = helper.clone();
-    helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::Flatten<#flatten_generic> },
-    );
-    let flatten_body = helper.map(
-            |fields| {
-                let mapped_fields = fields.iter().map(|MapField { ty, refer, .. }| quote! { <#ty as #substrate::types::Flatten<#flatten_generic>>::flatten(#refer, __substrate_output_sink); });
-                quote! { #(#mapped_fields)* }
-            },
-        );
-    helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::Flatten<#flatten_generic> },
-        trait_body: quote! {
-            fn flatten<E>(&self, __substrate_output_sink: &mut E)
-            where
-                E: ::std::iter::Extend<#flatten_generic> {
-                #flatten_body
-            }
-        },
-        extra_generics: vec![flatten_generic],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_unflatten(
-    kind_helper: &DeriveInputHelper,
-    view_helper: &DeriveInputHelper,
-    bundle_kind: &syn::Type,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let unflatten_generic = parse_quote! { SubstrateS };
-    let mut kind_helper = kind_helper.clone();
-    let mut view_helper = view_helper.clone();
-    view_helper.push_where_predicate_per_field(|_ty, prev_tys| {
-        let prev_ty = &prev_tys[0];
-        parse_quote! { #prev_ty: #substrate::types::HasBundleKind }
-    });
-    view_helper.push_where_predicate_per_field(
-        |ty, prev_tys| {
-            let prev_ty = &prev_tys[0];
-            parse_quote! { #ty: #substrate::types::Unflatten<<#prev_ty as #substrate::types::HasBundleKind>::BundleKind, #unflatten_generic> }
-        },
-    );
-    let unflatten_referent = quote! { __substrate_data };
-    kind_helper.set_referent(unflatten_referent.clone());
-    let unflatten_body = kind_helper.map_data(
-        &view_helper.get_type(),
-            |MapField { ty, refer, prev_tys, .. }| {
-                    let root_ty = if let Some(prev_ty) = prev_tys.first() {
-                        prev_ty
-                    } else { ty };
-                    quote! { <<#root_ty as #substrate::types::codegen::HasView<SubstrateV>>::View as #substrate::types::Unflatten<#ty, #unflatten_generic>>::unflatten(&#refer, __substrate_source)? }
-            });
-    view_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::Unflatten<#bundle_kind, #unflatten_generic> },
-        trait_body: quote! {
-            fn unflatten<SubstrateI>(#unflatten_referent: &#bundle_kind, __substrate_source: &mut SubstrateI) -> Option<Self>
-            where
-                SubstrateI: Iterator<Item = #unflatten_generic> {
-                ::std::option::Option::Some(#unflatten_body)
-            }
-        },
-        extra_generics: vec![unflatten_generic],
-        extra_where_predicates: vec![],
-    })
-}
-
-// fn impl_save(
-//     kind_helper: &DeriveInputHelper,
-//     view_helper: &DeriveInputHelper,
-//     bundle_kind: &syn::Type,
-// ) -> TokenStream {
-//     let substrate = substrate_ident();
-//     let simulator_ty = parse_quote! { __substrate_SIMULATOR };
-//     let analysis_ty = parse_quote! { __substrate_ANALYSIS };
-//     let mut view_helper = view_helper.clone();
-//     view_helper.push_where_predicate_per_field(|ty, _| {
-//         parse_quote! { #ty: #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> }
-//     });
-//     let view_ident = view_helper.get_ident();
-//     view_helper.impl_trait(&ImplTrait {
-//         trait_name: quote! { #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> },
-//         trait_body: quote! {
-//             type SaveKey = #view_ident<#substrate::types::codegen::NestedSaveKey<#hnv_generic_ty, __substrate_S, __substrate_A>>;
-//             type Saved = #view_ident<#substrate::types::codegen::NestedSaved<#hnv_generic_ty, __substrate_S, __substrate_A>>;
-//
-//             fn save(
-//                 &self,
-//                 ctx: &#substrate::simulation::SimulationContext<__substrate_S>,
-//                 opts: &mut <__substrate_S as #substrate::simulation::Simulator>::Options,
-//             ) -> <Self as #substrate::simulation::data::Save<__substrate_S, __substrate_A>>::SaveKey {
-//                 todo!()
-//             }
-//
-//             fn from_saved(
-//                 output: &<__substrate_A as #substrate::simulation::Analysis>::Output,
-//                 key: &<Self as #substrate::simulation::data::Save<__substrate_S, __substrate_A>>::SaveKey,
-//             ) -> <Self as #substrate::simulation::data::Save<__substrate_S, __substrate_A>>::Saved {
-//                 todo!()
-//             }
-//         },
-//         extra_generics: vec![simulator_ty, analysis_ty],
-//         extra_where_predicates: vec![],
-//     })
-// }
-
-fn impl_schematic_bundle_kind(
-    kind_helper: &DeriveInputHelper,
-    terminal_bundle_helper: &DeriveInputHelper,
-    node_bundle_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut schematic_bundle_kind_helper = kind_helper.clone();
-    schematic_bundle_kind_helper.push_where_predicate_per_field(|_ty, prev_tys| {
-        let prev_ty = &prev_tys[0];
-        parse_quote! { #prev_ty: #substrate::types::codegen::HasSchematicBundleKindViews }
-    });
-
-    let terminal_view_body = node_bundle_helper.double_map_data(
-        &terminal_bundle_helper.get_type(),
-        (&quote!{ cell_io }, &quote!{ instance_io }),
-            |MapField { refer: refer0, prev_tys, .. }, MapField { refer: refer1, .. }| {
-                let prev_ty = &prev_tys[0];
-                quote!{<<#prev_ty as #substrate::types::HasBundleKind>::BundleKind as #substrate::types::schematic::SchematicBundleKind>::terminal_view(cell, #refer0, instance, #refer1)}
-            }, quote!{ panic!("cell and instance IOs are not the same kind") });
-
-    schematic_bundle_kind_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::schematic::SchematicBundleKind },
-        trait_body: quote! {
-            fn terminal_view(
-                cell: #substrate::schematic::CellId,
-                cell_io: &#substrate::types::schematic::NodeBundle<Self>,
-                instance: #substrate::schematic::InstanceId,
-                instance_io: &#substrate::types::schematic::NodeBundle<Self>,
-            ) -> #substrate::types::schematic::TerminalBundle<Self> {
-                #terminal_view_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_has_node_bundle(
-    kind_helper: &DeriveInputHelper,
-    node_bundle_helper: &DeriveInputHelper,
-    terminal_bundle_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut kind_helper = kind_helper.clone();
-    kind_helper.push_where_predicate_per_field(|ty, _prev_tys| {
-        parse_quote! { #ty: #substrate::types::schematic::HasNodeBundle }
-    });
-    kind_helper.push_where_predicate_per_field(|ty, prev_tys| {
-        let prev_ty = if prev_tys.is_empty() {
-            ty
-        } else {
-            &prev_tys[0]
-        };
-        parse_quote! { #prev_ty: #substrate::types::codegen::HasSchematicBundleKindViews }
-    });
-
-    let node_bundle_full_ty = node_bundle_helper.get_full_type();
-
-    kind_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::schematic::HasNodeBundle },
-        trait_body: quote! {
-            type NodeBundle = #node_bundle_full_ty;
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_has_terminal_bundle(
-    kind_helper: &DeriveInputHelper,
-    node_bundle_helper: &DeriveInputHelper,
-    terminal_bundle_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut kind_helper = kind_helper.clone();
-    kind_helper.push_where_predicate_per_field(|ty, _prev_tys| {
-        parse_quote! { #ty: #substrate::types::schematic::HasTerminalBundle }
-    });
-    kind_helper.push_where_predicate_per_field(|ty, prev_tys| {
-        let prev_ty = if prev_tys.is_empty() {
-            ty
-        } else {
-            &prev_tys[0]
-        };
-        parse_quote! { #prev_ty: #substrate::types::codegen::HasSchematicBundleKindViews }
-    });
-
-    let terminal_bundle_full_ty = terminal_bundle_helper.get_full_type();
-
-    kind_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::types::schematic::HasTerminalBundle },
-        trait_body: quote! {
-            type TerminalBundle = #terminal_bundle_full_ty;
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_save_nested_bundle(view_helper: &DeriveInputHelper, nodes: bool) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut view_helper = view_helper.clone();
-    let simulator_ty = parse_quote! { SubstrateS };
-    let analysis_ty = parse_quote! { SubstrateA };
-
-    let (save_key_view, saved_view, bundle_view) = if nodes {
-        (
-            parse_quote! { #substrate::types::codegen::NestedNodeSaveKeyView<#simulator_ty, #analysis_ty> },
-            parse_quote! { #substrate::types::codegen::NestedNodeSavedView<#simulator_ty, #analysis_ty> },
-            parse_quote! { #substrate::types::codegen::NestedNodeBundle },
-        )
-    } else {
-        (
-            parse_quote! { #substrate::types::codegen::NestedTerminalSaveKeyView<#simulator_ty, #analysis_ty> },
-            parse_quote! { #substrate::types::codegen::NestedTerminalSavedView<#simulator_ty, #analysis_ty> },
-            parse_quote! { #substrate::types::codegen::NestedTerminalBundle },
-        )
-    };
-    let mut save_key = view_helper.clone();
-    save_key.add_generic_type_binding(parse_quote! { SubstrateV }, save_key_view);
-    let mut saved = view_helper.clone();
-    saved.add_generic_type_binding(parse_quote! { SubstrateV }, saved_view);
-
-    view_helper.push_where_predicate_per_field(|ty, _prev_tys| {
-        parse_quote! { #ty: #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> }
-    });
-
-    view_helper.add_generic_type_binding(parse_quote! { SubstrateV }, bundle_view);
-
-    let save_body = view_helper.map_data(
-        &save_key.get_full_turbofish_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::save(&#refer, __substrate_ctx, __substrate_opts) }
-            });
-    let mut from_saved_helper = view_helper.clone();
-    from_saved_helper.set_referent(quote! { __substrate_key });
-    let from_saved_body = from_saved_helper.map_data(
-        &saved.get_full_turbofish_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::from_saved(__substrate_output, #refer) }
-            });
-
-    let save_key_full_ty = save_key.get_full_type();
-    let saved_full_ty = saved.get_full_type();
-
-    view_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> },
-        trait_body: quote! {
-            type SaveKey = #save_key_full_ty;
-            type Saved = #saved_full_ty;
-            fn save(
-                &self,
-                __substrate_ctx: &#substrate::simulation::SimulationContext<#simulator_ty>,
-                __substrate_opts: &mut <#simulator_ty as #substrate::simulation::Simulator>::Options,
-            ) -> <Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::SaveKey {
-                #save_body
-            }
-
-            fn from_saved(
-                __substrate_output: &<#analysis_ty as #substrate::simulation::Analysis>::Output,
-                __substrate_key: &<Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::SaveKey,
-            ) -> <Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::Saved {
-                #from_saved_body
-            }
-        },
-        extra_where_predicates: vec![
-            parse_quote! { #simulator_ty: #substrate::simulation::Simulator },
-            parse_quote! { #analysis_ty: #substrate::simulation::Analysis },
-        ],
-        extra_generics: vec![simulator_ty, analysis_ty],
-    })
-}
-
-fn impl_save_nested_native(view_helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut view_helper = view_helper.clone();
-    let simulator_ty = parse_quote! { SubstrateS };
-    let analysis_ty = parse_quote! { SubstrateA };
-
-    let hnv_generic_ty: syn::Ident = parse_quote!(SubstrateT);
-    let hnv_generic: syn::GenericParam = parse_quote!(#hnv_generic_ty);
-
-    let save_key_view = parse_quote! { #substrate::types::codegen::NestedSaveKey<#hnv_generic_ty, #simulator_ty, #analysis_ty> };
-    let saved_view = parse_quote! { #substrate::types::codegen::NestedSaved<#hnv_generic_ty, #simulator_ty, #analysis_ty> };
-
-    let mut save_key = view_helper.clone();
-    save_key.add_generic_type_binding(parse_quote! { SubstrateV }, save_key_view);
-    let mut saved = view_helper.clone();
-    saved.add_generic_type_binding(parse_quote! { SubstrateV }, saved_view);
-
-    view_helper.push_where_predicate_per_field(|ty, prev_tys| {
-        let ty = if prev_tys.is_empty() {
-            ty
-        } else {
-            &prev_tys[0]
-        };
-        parse_quote! { #ty: #substrate::schematic::HasNestedView<#hnv_generic_ty> }
-    });
-    view_helper.push_where_predicate_per_field(|ty, prev_tys| {
-        let ty = if prev_tys.is_empty() {
-            ty
-        } else {
-            &prev_tys[0]
-        };
-        parse_quote! {<#ty as #substrate::schematic::HasNestedView<#hnv_generic_ty>>::NestedView: #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> }
-    });
-
-    view_helper.add_generic_type_binding(
-        parse_quote! { SubstrateV },
-        parse_quote! { #substrate::types::codegen::Nested<#hnv_generic_ty> },
-    );
-
-    let save_body = view_helper.map_data(
-        &save_key.get_full_turbofish_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::save(&#refer, __substrate_ctx, __substrate_opts) }
-            });
-    let mut from_saved_helper = view_helper.clone();
-    from_saved_helper.set_referent(quote! { __substrate_key });
-    let from_saved_body = from_saved_helper.map_data(
-        &saved.get_full_turbofish_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::from_saved(__substrate_output, #refer) }
-            });
-
-    let save_key_full_ty = save_key.get_full_type();
-    let saved_full_ty = saved.get_full_type();
-
-    view_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::simulation::data::Save<#simulator_ty, #analysis_ty> },
-        trait_body: quote! {
-            type SaveKey = #save_key_full_ty;
-            type Saved = #saved_full_ty;
-            fn save(
-                &self,
-                __substrate_ctx: &#substrate::simulation::SimulationContext<#simulator_ty>,
-                __substrate_opts: &mut <#simulator_ty as #substrate::simulation::Simulator>::Options,
-            ) -> <Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::SaveKey {
-                #save_body
-            }
-
-            fn from_saved(
-                __substrate_output: &<#analysis_ty as #substrate::simulation::Analysis>::Output,
-                __substrate_key: &<Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::SaveKey,
-            ) -> <Self as #substrate::simulation::data::Save<#simulator_ty, #analysis_ty>>::Saved {
-                #from_saved_body
-            }
-        },
-        extra_where_predicates: vec![
-            parse_quote! { #simulator_ty: #substrate::simulation::Simulator },
-            parse_quote! { #analysis_ty: #substrate::simulation::Analysis },
-        ],
-        extra_generics: vec![hnv_generic, simulator_ty, analysis_ty],
-    })
-}
-
-fn impl_has_nested_view(
-    view_helper: &DeriveInputHelper,
-    nested_view_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut has_nested_view_helper = view_helper.clone();
-    has_nested_view_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::schematic::HasNestedView },
-    );
-
-    let nested_view_full_ty = nested_view_helper.get_full_type();
-
-    let nested_view_body = has_nested_view_helper.map_data(
-        &nested_view_helper.get_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::schematic::HasNestedView>::nested_view(&#refer, __substrate_parent) }
-            });
-    has_nested_view_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::schematic::HasNestedView },
-        trait_body: quote! {
-            type NestedView = #nested_view_full_ty;
-
-            fn nested_view(&self, __substrate_parent: &#substrate::schematic::InstancePath) -> #substrate::schematic::NestedView<Self> {
-                #nested_view_body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_view_as(view_helper: &DeriveInputHelper, nodes: bool) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut view_as_helper = view_helper.clone();
-
-    let (bundle_view_ident, view_as_fn) = if nodes {
-        (quote! { NodeBundle }, quote! { view_nodes_as })
-    } else {
-        (quote! { TerminalBundle }, quote! { view_terminals_as })
-    };
-    view_as_helper.push_where_predicate(parse_quote!{ Self: #substrate::types::HasBundleKind<BundleKind: #substrate::types::schematic::SchematicBundleKind<#bundle_view_ident = Self>> });
-
-    let full_ty = view_as_helper.get_full_type();
-    let (imp, _, wher) = view_as_helper.custom_split_for_impl();
-    let vis = &view_as_helper.get_input().vis;
-
-    quote! {
-        impl #imp #full_ty #wher {
-            /// Views this node bundle as a node bundle of a different kind.
-            #vis fn view_as<SubstrateT: #substrate::types::HasBundleKind<BundleKind: #substrate::types::schematic::SchematicBundleKind>>(&self) -> #substrate::types::schematic::#bundle_view_ident<<SubstrateT as #substrate::types::HasBundleKind>::BundleKind> where <Self as #substrate::types::HasBundleKind>::BundleKind: #substrate::types::schematic::DataView<<SubstrateT as #substrate::types::HasBundleKind>::BundleKind>{
-                <<Self as #substrate::types::HasBundleKind>::BundleKind as #substrate::types::schematic::DataView<<SubstrateT as #substrate::types::HasBundleKind>::BundleKind>>::#view_as_fn(self)
-            }
-        }
-    }
-}
-
-fn impl_translate_ref(view_helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut translate_ref_helper = view_helper.clone();
-    translate_ref_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::geometry::transform::TranslateRef },
-    );
-
-    let body = translate_ref_helper.map_data(
-        &translate_ref_helper.get_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::geometry::transform::TranslateRef>::translate_ref(&#refer, __substrate_point) }
-            });
-    translate_ref_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::geometry::transform::TranslateRef },
-        trait_body: quote! {
-            fn translate_ref(&self, __substrate_point: #substrate::geometry::point::Point) -> Self {
-                #body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
-
-fn impl_transform_ref(view_helper: &DeriveInputHelper) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut transform_ref_helper = view_helper.clone();
-    transform_ref_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::geometry::transform::TransformRef },
-    );
-
-    let body = transform_ref_helper.map_data(
-        &transform_ref_helper.get_type(),
-            |MapField { ty, refer, .. }| {
-                    quote! { <#ty as #substrate::geometry::transform::TransformRef>::transform_ref(&#refer, __substrate_transformation) }
-            });
-    transform_ref_helper.impl_trait(&ImplTrait {
-        trait_name: quote! { #substrate::geometry::transform::TransformRef },
-        trait_body: quote! {
-            fn transform_ref(&self, __substrate_transformation: #substrate::geometry::transform::Transformation) -> Self {
-                #body
-            }
-        },
-        extra_generics: vec![],
-        extra_where_predicates: vec![],
-    })
-}
 
 /// Derives `BundleKind` for the provided input and creates a struct representing views of this
 /// `BundleKind`.
@@ -789,168 +97,4 @@ pub(crate) fn bundle_kind(input: &DeriveInput, io: bool) -> syn::Result<TokenStr
     Ok(quote! {
         #( #all_decls_impls )*
     })
-}
-
-/// Derives `HasNestedView` and `Save` for the input.
-pub(crate) fn nested_data(input: &DeriveInput) -> syn::Result<TokenStream> {
-    let substrate = substrate_ident();
-    let helper = DeriveInputHelper::new(input.clone())?;
-    let view_ident = format_ident!("{}View", &input.ident);
-    let mut all_decls_impls = Vec::new();
-
-    // Create `View` struct
-    // - Needs to add a generic along with a where clause per field that uses that generic
-    // - Potentially be able to add separate where clauses to new generic
-    let mut view_helper = helper.clone();
-    view_helper.set_ident(view_ident);
-    let view_generic_ty = quote! { SubstrateV };
-    view_helper.push_generic_param(parse_quote! { #view_generic_ty });
-    let mut save_helper = view_helper.clone();
-    // These where clauses should only be pushed to the view helper, not the save helper.
-    // [`impl_save_nested_native`] will add sufficient where clauses, and adding the where
-    // clauses below seems to confuse the trait solver.
-    view_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::codegen::HasView<#view_generic_ty> },
-    );
-    for helper in [&mut view_helper, &mut save_helper] {
-        helper.map_types(
-            |ty| parse_quote! { <#ty as #substrate::types::codegen::HasView<#view_generic_ty>>::View },
-        );
-    }
-    all_decls_impls.push(view_helper.decl_data());
-    all_decls_impls.push(impl_flatlen(&view_helper));
-    all_decls_impls.push(impl_flatten_generic(&view_helper));
-
-    let mut hnv_helper = view_helper.clone();
-    hnv_helper.add_generic_type_binding(
-        parse_quote!(#view_generic_ty),
-        parse_quote!(#substrate::types::codegen::Nested),
-    );
-
-    all_decls_impls.push(impl_has_nested_view(&helper, &hnv_helper));
-    all_decls_impls.push(impl_save_nested_native(&save_helper));
-    Ok(quote! {
-        #( #all_decls_impls )*
-    })
-}
-
-pub(crate) fn schematic_bundle_kind(
-    original_helper: &DeriveInputHelper,
-    kind_helper: &DeriveInputHelper,
-    view_helper: &DeriveInputHelper,
-) -> TokenStream {
-    let substrate = substrate_ident();
-    let mut all_decls_impls = Vec::new();
-    let view_generic_ty = quote! { SubstrateV };
-
-    let helper = original_helper.clone();
-    let mut node_bundle_helper = original_helper.clone();
-    node_bundle_helper.set_ident(view_helper.get_ident().clone());
-    node_bundle_helper.push_generic_param(parse_quote! { #view_generic_ty });
-    node_bundle_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::types::codegen::HasSchematicBundleKindViews },
-    );
-
-    let mut terminal_bundle_helper = node_bundle_helper.clone();
-    let mut nested_node_bundle_helper = node_bundle_helper.clone();
-    let mut nested_terminal_bundle_helper = node_bundle_helper.clone();
-
-    node_bundle_helper.add_generic_type_binding(
-        parse_quote! { #view_generic_ty },
-        parse_quote! { #substrate::types::codegen::NodeBundle },
-    );
-    node_bundle_helper.map_types(
-        |ty| parse_quote! { <#ty as #substrate::types::codegen::HasView<#substrate::types::codegen::NodeBundle>>::View },
-    );
-
-    terminal_bundle_helper.add_generic_type_binding(
-        parse_quote! { #view_generic_ty },
-        parse_quote! { #substrate::types::codegen::TerminalBundle },
-    );
-    terminal_bundle_helper.map_types(
-        |ty| parse_quote! { <#ty as #substrate::types::codegen::HasView<#substrate::types::codegen::TerminalBundle>>::View },
-    );
-
-    nested_node_bundle_helper.add_generic_type_binding(
-        parse_quote! { #view_generic_ty },
-        parse_quote! { #substrate::types::codegen::NestedNodeBundle },
-    );
-    nested_node_bundle_helper.map_types(
-        |ty| parse_quote! { <#ty as #substrate::types::codegen::HasView<#substrate::types::codegen::NestedNodeBundle>>::View },
-    );
-    nested_node_bundle_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::schematic::HasNestedView<NestedView = #ty> },
-    );
-
-    nested_terminal_bundle_helper.add_generic_type_binding(
-        parse_quote! { #view_generic_ty },
-        parse_quote! { #substrate::types::codegen::NestedTerminalBundle },
-    );
-    nested_terminal_bundle_helper.map_types(
-        |ty| parse_quote! { <#ty as #substrate::types::codegen::HasView<#substrate::types::codegen::NestedTerminalBundle>>::View },
-    );
-    nested_terminal_bundle_helper.push_where_predicate_per_field(
-        |ty, _| parse_quote! { #ty: #substrate::schematic::HasNestedView<NestedView = #ty>},
-    );
-
-    all_decls_impls.push(impl_has_node_bundle(
-        &helper,
-        &node_bundle_helper,
-        &terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_terminal_bundle(
-        &helper,
-        &node_bundle_helper,
-        &terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_node_bundle(
-        kind_helper,
-        &node_bundle_helper,
-        &terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_terminal_bundle(
-        kind_helper,
-        &node_bundle_helper,
-        &terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_schematic_bundle_kind(
-        kind_helper,
-        &node_bundle_helper,
-        &terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_nested_view(
-        &node_bundle_helper,
-        &nested_node_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_nested_view(
-        &terminal_bundle_helper,
-        &nested_terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_nested_view(
-        &nested_node_bundle_helper,
-        &nested_node_bundle_helper,
-    ));
-    all_decls_impls.push(impl_has_nested_view(
-        &nested_terminal_bundle_helper,
-        &nested_terminal_bundle_helper,
-    ));
-    all_decls_impls.push(impl_view_as(&node_bundle_helper, true));
-    all_decls_impls.push(impl_view_as(&terminal_bundle_helper, false));
-    all_decls_impls.push(impl_save_nested_bundle(view_helper, true));
-    all_decls_impls.push(impl_save_nested_bundle(view_helper, false));
-
-    quote! {
-        #( #all_decls_impls )*
-    }
-}
-
-pub(crate) fn layout_bundle_kind(view_helper: &DeriveInputHelper) -> TokenStream {
-    let mut all_decls_impls = Vec::new();
-
-    all_decls_impls.push(impl_translate_ref(view_helper));
-    all_decls_impls.push(impl_transform_ref(view_helper));
-
-    quote! {
-        #( #all_decls_impls )*
-    }
 }
